requirements.txt
Django==1.6.5
Jinja2==2.7.3
Markdown==2.4.1
MarkupSafe==0.23
MySQL-python==1.2.5
Pygments==1.6
South==0.8.4
Sphinx==1.2.2
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.18
celery==3.1.12
coverage==3.7.1
django-celery==3.1.10
django-debug-toolbar==1.2.1
django-filter==0.7
django-model-utils==2.0.3
django-pandas==0.2.0
django-registration==1.0
django-registration-defaults==0.4.4
django-tastypie==0.11.1
djangorestframework==2.3.14
docutils==0.11
kombu==3.0.20
numpy==1.8.1
pandas==0.14.1
py==1.4.23
pycrypto==2.6.1
pytest==2.6.1
pytest-django==2.6.2
python-dateutil==2.2
python-mimeparse==0.1.4
pytz==2014.4
redis==2.10.1
rest-pandas==0.1.1
selenium==2.42.1
six==1.7.2
sqlparse==0.1.11
wsgiref==0.1.2
